+++ changelog.md
@@ -22,7 +22,8 @@ This version comes with a 1.8.9 port in addition to many additions.
   * [1.7.10] ADDED: getPosition() to tape drives. No, I'm not joking. It's real. I'm sorry.
   * [1.7.10] FIXED: Crash when the `openComputersBees` config option was set to `false`.
   * [1.7.10] FIXED: Digital Locomotive Relay trying to consume OpenComputers energy even if called from ComputerCraft
-  * FIXED: Updated OpenComputers integration
+  * [1.7.10] FIXED: Updated OpenComputers integration
+  * [1.7.10] FIXED: Removed explicit dependency on GregTech 5. Computronics will launch with GregTech 6 now (but there is no integration yet).
 
   * [1.8.9] Ported the mod to 1.8.9! Everything is resource-pack ready!
   	- Mods supported right now are ComputerCraft, OpenComputers, TIS-3D, Flamingo, Storage Drawers, Charset and WAILA.